rtf_orchestrator::db

Trait StatusTracked

pub trait StatusTracked: Queryable {
    const STATUS_TABLE: &'static str;

    // Required method
    fn after_set_status(
        &self,
        status: Status,
        conn: &mut PgConnection,
    ) -> impl Future<Output = Result<()>> + Send;

    // Provided methods
    fn set_status(
        &self,
        status: Status,
        message: Option<String>,
        conn: &mut PgConnection,
    ) -> impl Future<Output = Result<()>> + Send { ... }
    fn try_current_status(
        &self,
        conn: &mut PgConnection,
    ) -> impl Future<Output = Result<Option<StatusUpdate>>> + Send { ... }
    fn current_status(
        &self,
        conn: &mut PgConnection,
    ) -> impl Future<Output = Result<StatusUpdate>> + Send { ... }
    fn status_history(
        &self,
        conn: &mut PgConnection,
    ) -> impl Future<Output = Result<Vec<StatusUpdate>>> + Send { ... }
}

Helper trait for tracking a time series of StatusUpdate items for a parent table.

Primary table requirements

Must contain a nullable timestamp “completed_at” column

Status table structure

This trait requires a fixed structure for the status table:

integer “parent_id”

integer “status”

nullable text “message”

timestamp “updated_at”

Semantics

See the documentation on the Status enum for how each of the different statuses are used.

Required Associated Constants

const STATUS_TABLE: &'static str

Required Methods

fn after_set_status( &self, status: Status, conn: &mut PgConnection, ) -> impl Future<Output = Result<()>> + Send

Additional logic to run after recording a status update for this type.

Provided Methods

fn set_status( &self, status: Status, message: Option<String>, conn: &mut PgConnection, ) -> impl Future<Output = Result<()>> + Send

fn try_current_status( &self, conn: &mut PgConnection, ) -> impl Future<Output = Result<Option<StatusUpdate>>> + Send

fn current_status( &self, conn: &mut PgConnection, ) -> impl Future<Output = Result<StatusUpdate>> + Send

fn status_history( &self, conn: &mut PgConnection, ) -> impl Future<Output = Result<Vec<StatusUpdate>>> + Send

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.