rtf_orchestrator::endpoints

Module admin

Admin routes are only accessible to users in the Orchestrator’s “admins” configmap.

See: https://github.com/mdg-private/kanaveral/blob/main/management-cluster/orchestrator/admins.yaml

Modules

known_test_plan_cluster_pin

Pin or unpin the workload cluster a known test plan’s runs execute in.

purge_queue

Queue management for purging known “bad” / stuck runs and executions.

register_known_test_plan

Register a Test Plan with the orchestrator so it can be triggered and queried by UUID or name.

Functions

event_queue_snapshot_handler

get_logging_filter_handler

set_logging_filter_handler