rtf_orchestrator::resolver

Function resolver_task

pub async fn resolver_task(
    rx: UnboundedReceiver<ResolverInput>,
    prov_handle: ProvisioningHandle,
) -> Result<()>

A long lived Tokio task that is responsible for running all RTF related logic that executes within the server.

See resolve_variant below for the specific rtf_config logic that is run on the server.

Differences compared to rtf_cli

The resolution logic used here only supports processing a OrchestratorTestPlan that has been submitted as part of a PreparedPayload (prepared using rtf remote prepare on the command line). That preparation logic handles all filesystem operations on the client side and provides the required local file data for us to construct a OrchestratorContext that can then handle resolving what’s left.